refactor: turn TransmodelRequestContext into a request-scope-backed interface
Same motivation as the GTFS context: the class's constructor forced every
dependency to be resolved as soon as any Transmodel data fetcher touched the
context. TransmodelRequestContext is now an interface; the production
implementation delegates each getter to the enclosing RequestScopedFactory,
so a dependency is only resolved on first actual use. Tests that need a
plain, hand-built context use the new TestTransmodelRequestContext class.

# Conflicts:
#	application/src/test/java/org/opentripplanner/apis/transmodel/mapping/TripRequestMapperTest.java
